refactor(core): extract error-boundary to own directive file
Move error-boundary directive from validation.js into dedicated
error-boundary.js. The directive stays in core — only validate
will be extracted to Elements later. Update test imports accordingly.
